pyproject.toml
[project]
name = "sahaf"
version = "0.1.0"
description = "PDF to Markdown converter with OCR support — local, fast, accurate."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pymupdf>=1.24.0",
    "marker-pdf>=1.0.0",
    "ebooklib>=0.18",
    "markdownify>=0.11.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
